Moving off of IRC primarily
Moving off of IRC to make for a better contributor initial experience.
IRC is great, we love it, but it's not the most approachable for the
incoming contributors.  Moving to Matrix as it's got persistent clients,
ability to read scroll-back, search, and generally gets information out
slightly better.

This will also coincide with shutting the slack and eventual irc bridges
down entirely.

Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>